@@ -14,29 +14,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
+    "@angular/animations": "^18.2.13",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
     "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
     "@ckeditor/ckeditor5-angular": "^7.0.1",
     "@ckeditor/ckeditor5-build-classic": "^41.1.0",
     "@ckeditor/ckeditor5-core": "^41.1.0",
     "@ckeditor/ckeditor5-engine": "^41.1.0",
     "@ckeditor/ckeditor5-utils": "^41.1.0",
     "@ckeditor/ckeditor5-watchdog": "^41.1.0",
-    "angular-gridster2": "^17.0.0",
+    "angular-gridster2": "^18.0.1",
     "browserify-fs": "^1.0.0",
     "browserify-zlib": "^0.2.0",
     "chart.js": "^4.4.1",
     "jdenticon": "^3.2.0",
     "ng2-charts": "^4.1.1",
-    "ngx-cookie-service": "^17.1.0",
+    "ngx-cookie-service": "^18.0.0",
     "ngx-jdenticon": "^2.0.0",
     "ngx-ui-scroll": "^3.2.0-beta.2",
     "node-polyfill-webpack-plugin": "^2.0.1",
@@ -49,15 +49,15 @@
     "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "17.0.2",
-    "@angular-devkit/build-angular": "^17.3.17",
-    "@angular-eslint/builder": "17.5.3",
-    "@angular-eslint/eslint-plugin": "17.5.3",
-    "@angular-eslint/eslint-plugin-template": "17.5.3",
-    "@angular-eslint/schematics": "17.5.3",
-    "@angular-eslint/template-parser": "17.5.3",
-    "@angular/cli": "^17.3.17",
-    "@angular/compiler-cli": "^17.3.12",
+    "@angular-builders/custom-webpack": "18.0.0",
+    "@angular-devkit/build-angular": "^18.2.19",
+    "@angular-eslint/builder": "18.4.3",
+    "@angular-eslint/eslint-plugin": "18.4.3",
+    "@angular-eslint/eslint-plugin-template": "18.4.3",
+    "@angular-eslint/schematics": "18.4.3",
+    "@angular-eslint/template-parser": "18.4.3",
+    "@angular/cli": "^18.2.19",
+    "@angular/compiler-cli": "^18.2.13",
     "@scicatproject/scicat-sdk-ts-fetch": "^4.13.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
